Remove dependency on zumba/json-serializer
Instead of deserializing the data and then comparing it with the actual
objects, this serializes the actual object and then compares it with the
serialized data.

Adds duplicated information and the type of objects are lost, but this
allows for the removal of zumba/json-serializer dependency.

We can develop a specialized serializer for this or use another method
to test this data later on.

Signed-off-by: Maurício Meneghini Fauth <mauricio@mfauth.net>